Refactored usm_to_string method to usm_common.h
